+++ servant-client-core/src/Servant/Client/Core/Internal/BasicAuth.hs
@@ -6,11 +6,16 @@
 
 module Servant.Client.Core.Internal.BasicAuth where
 
-import           Data.ByteString.Base64               (encode)
-import           Data.Monoid                          ((<>))
-import           Data.Text.Encoding                   (decodeUtf8)
-import           Servant.API.BasicAuth                (BasicAuthData (BasicAuthData))
-import           Servant.Client.Core.Internal.Request (Request, addHeader)
+import           Data.ByteString.Base64
+                 (encode)
+import           Data.Monoid
+                 ((<>))
+import           Data.Text.Encoding
+                 (decodeUtf8)
+import           Servant.API.BasicAuth
+                 (BasicAuthData (BasicAuthData))
+import           Servant.Client.Core.Internal.Request
+                 (Request, addHeader)
 
 -- | Authenticate a request using Basic Authentication
 basicAuthReq :: BasicAuthData -> Request -> Request