+++ servant-client-core/src/Servant/Client/Free.hs
@@ -1,15 +1,18 @@
-{-# LANGUAGE ScopedTypeVariables, FlexibleContexts, GADTs #-}
+{-# LANGUAGE FlexibleContexts    #-}
+{-# LANGUAGE GADTs               #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 module Servant.Client.Free (
     client,
     ClientF (..),
     module Servant.Client.Core.Reexport,
     ) where
 
-import Data.Proxy (Proxy (..))
-import Control.Monad.Free
-import Servant.Client.Core
-import Servant.Client.Core.Reexport
-import Servant.Client.Core.Internal.ClientF
+import           Control.Monad.Free
+import           Data.Proxy
+                 (Proxy (..))
+import           Servant.Client.Core
+import           Servant.Client.Core.Internal.ClientF
+import           Servant.Client.Core.Reexport
 
 client :: HasClient (Free ClientF) api => Proxy api -> Client (Free ClientF) api
 client api = api `clientIn` (Proxy :: Proxy (Free ClientF))