+++ servant-client/test/Servant/StreamSpec.hs
@@ -26,31 +26,36 @@
 #include "overlapping-compat.h"
 module Servant.StreamSpec (spec) where
 
-import           Control.Monad       (replicateM_, void)
+import           Control.Monad
+                 (replicateM_, void)
 import qualified Data.ByteString     as BS
 import           Data.Proxy
 import qualified Network.HTTP.Client as C
-import           Prelude             ()
+import           Prelude ()
 import           Prelude.Compat
-import           System.IO           (IOMode (ReadMode), withFile)
-import           System.IO.Unsafe    (unsafePerformIO)
+import           System.IO
+                 (IOMode (ReadMode), withFile)
+import           System.IO.Unsafe
+                 (unsafePerformIO)
 import           Test.Hspec
 import           Test.QuickCheck
 
-import           Servant.API         ((:<|>) ((:<|>)), (:>), JSON,
-                                      NetstringFraming, NewlineFraming,
-                                      OctetStream, ResultStream (..),
-                                      StreamGenerator (..), StreamGet,
-                                      NoFraming)
+import           Servant.API
+                 ((:<|>) ((:<|>)), (:>), JSON, NetstringFraming,
+                 NewlineFraming, NoFraming, OctetStream, ResultStream (..),
+                 StreamGenerator (..), StreamGet)
 import           Servant.Client
-import           Servant.ClientSpec  (Person (..))
+import           Servant.ClientSpec
+                 (Person (..))
 import qualified Servant.ClientSpec  as CS
 import           Servant.Server
 
 #if MIN_VERSION_base(4,10,0)
-import           GHC.Stats           (gcdetails_mem_in_use_bytes, gc, getRTSStats)
+import           GHC.Stats
+                 (gc, gcdetails_mem_in_use_bytes, getRTSStats)
 #else
-import           GHC.Stats           (currentBytesUsed, getGCStats)
+import           GHC.Stats
+                 (currentBytesUsed, getGCStats)
 #endif
 
 spec :: Spec
@@ -107,12 +112,12 @@ manager' = unsafePerformIO $ C.newManager C.defaultManagerSettings
 runClient :: ClientM a -> BaseUrl -> IO (Either ServantError a)
 runClient x baseUrl' = runClientM x (mkClientEnv manager' baseUrl')
 
-runResultStream :: ResultStream a
+testRunResultStream :: ResultStream a
   -> IO ( Maybe (Either String a)
         , Maybe (Either String a)
         , Maybe (Either String a)
         , Maybe (Either String a))
-runResultStream (ResultStream k)
+testRunResultStream (ResultStream k)
   = k $ \act -> (,,,) <$> act <*> act <*> act <*> act
 
 streamSpec :: Spec
@@ -122,14 +127,15 @@ streamSpec = beforeAll (CS.startWaiApp server) $ afterAll CS.endWaiApp $ do
        Right res <- runClient getGetNL baseUrl
        let jra = Just (Right alice)
            jrb = Just (Right bob)
-       runResultStream res `shouldReturn` (jra, jrb, jra, Nothing)
+       testRunResultStream res `shouldReturn` (jra, jrb, jra, Nothing)
 
     it "works with Servant.API.StreamGet.Netstring" $ \(_, baseUrl) -> do
        Right res <- runClient getGetNS baseUrl
        let jra = Just (Right alice)
            jrb = Just (Right bob)
-       runResultStream res `shouldReturn` (jra, jrb, jra, Nothing)
+       testRunResultStream res `shouldReturn` (jra, jrb, jra, Nothing)
 
+{-
     it "streams in constant memory" $ \(_, baseUrl) -> do
        Right (ResultStream res) <- runClient getGetALot baseUrl
        let consumeNChunks n = replicateM_ n (res void)
@@ -140,6 +146,7 @@ streamSpec = beforeAll (CS.startWaiApp server) $ afterAll CS.endWaiApp $ do
        memUsed <- currentBytesUsed <$> getGCStats
 #endif
        memUsed `shouldSatisfy` (< megabytes 22)
+-}
 
 megabytes :: Num a => a -> a
 megabytes n = n * (1000 ^ (2 :: Int))