+++ servant-server/src/Servant/Server/Internal/BasicAuth.hs
@@ -1,22 +1,31 @@
 {-# LANGUAGE DeriveDataTypeable #-}
-{-# LANGUAGE DeriveFunctor #-}
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE DeriveFunctor      #-}
+{-# LANGUAGE DeriveGeneric      #-}
+{-# LANGUAGE OverloadedStrings  #-}
 
 module Servant.Server.Internal.BasicAuth where
 
-import           Control.Monad          (guard)
-import           Control.Monad.Trans    (liftIO)
-import qualified Data.ByteString        as BS
-import           Data.ByteString.Base64 (decodeLenient)
-import           Data.Monoid            ((<>))
-import           Data.Typeable          (Typeable)
-import           Data.Word8             (isSpace, toLower, _colon)
+import           Control.Monad
+                 (guard)
+import           Control.Monad.Trans
+                 (liftIO)
+import qualified Data.ByteString                            as BS
+import           Data.ByteString.Base64
+                 (decodeLenient)
+import           Data.Monoid
+                 ((<>))
+import           Data.Typeable
+                 (Typeable)
+import           Data.Word8
+                 (isSpace, toLower, _colon)
 import           GHC.Generics
-import           Network.HTTP.Types     (Header)
-import           Network.Wai            (Request, requestHeaders)
+import           Network.HTTP.Types
+                 (Header)
+import           Network.Wai
+                 (Request, requestHeaders)
 
-import           Servant.API.BasicAuth (BasicAuthData(BasicAuthData))
+import           Servant.API.BasicAuth
+                 (BasicAuthData (BasicAuthData))
 import           Servant.Server.Internal.RoutingApplication
 import           Servant.Server.Internal.ServantErr