+++ servant-server/src/Servant/Server/Internal/RoutingApplication.hs
@@ -10,17 +10,27 @@
 {-# LANGUAGE UndecidableInstances       #-}
 module Servant.Server.Internal.RoutingApplication where
 
-import           Control.DeepSeq                    (force)
-import           Control.Monad                      (ap, liftM)
-import           Control.Monad.Base                 (MonadBase (..))
-import           Control.Monad.Catch                (MonadThrow (..))
-import           Control.Monad.Reader               (MonadReader (..), ReaderT (..), runReaderT)
-import           Control.Monad.Trans                (MonadIO (..), MonadTrans (..))
-import           Control.Monad.Trans.Control        (ComposeSt, MonadBaseControl (..), MonadTransControl (..),
-                                                     defaultLiftBaseWith, defaultRestoreM)
-import           Control.Monad.Trans.Resource       (MonadResource (..), ResourceT, runResourceT, transResourceT, withInternalState, runInternalState)
-import           Network.Wai                        (Application, Request, Response, ResponseReceived)
-import           Prelude                            ()
+import           Control.DeepSeq
+                 (force)
+import           Control.Monad
+                 (ap, liftM)
+import           Control.Monad.Base
+                 (MonadBase (..))
+import           Control.Monad.Catch
+                 (MonadThrow (..))
+import           Control.Monad.Reader
+                 (MonadReader (..), ReaderT (..), runReaderT)
+import           Control.Monad.Trans
+                 (MonadIO (..), MonadTrans (..))
+import           Control.Monad.Trans.Control
+                 (ComposeSt, MonadBaseControl (..), MonadTransControl (..),
+                 defaultLiftBaseWith, defaultRestoreM)
+import           Control.Monad.Trans.Resource
+                 (MonadResource (..), ResourceT, runInternalState,
+                 runResourceT, transResourceT, withInternalState)
+import           Network.Wai
+                 (Application, Request, Response, ResponseReceived)
+import           Prelude ()
 import           Prelude.Compat
 import           Servant.Server.Internal.Handler
 import           Servant.Server.Internal.ServantErr
@@ -94,7 +104,7 @@ toApplication fullyEvaluate ra request respond =
     maybeEval resp =
       if fullyEvaluate
         then force resp
-        else resp 
+        else resp
     routingRespond :: RouteResult Response -> IO ResponseReceived
     routingRespond (Fail err)      = respond $ responseServantErr err
     routingRespond (FailFatal err) = respond $ responseServantErr err