+++ servant-server/test/Servant/ArbitraryMonadServerSpec.hs
@@ -9,9 +9,10 @@ import           Data.Proxy
 import           Servant.API
 import           Servant.Server
 
-import           Test.Hspec                 (Spec, describe, it)
-import           Test.Hspec.Wai             (get, matchStatus, post,
-                                             shouldRespondWith, with)
+import           Test.Hspec
+                 (Spec, describe, it)
+import           Test.Hspec.Wai
+                 (get, matchStatus, post, shouldRespondWith, with)
 
 spec :: Spec
 spec = describe "Arbitrary monad server" $ do