+++ servant/src/Servant/API/Internal/Test/ComprehensiveAPI.hs
@@ -7,6 +7,7 @@
 module Servant.API.Internal.Test.ComprehensiveAPI where
 
 import           Data.Proxy
+                 (Proxy (..))
 import           Servant.API
 
 type GET = Get '[JSON] NoContent
@@ -38,6 +39,7 @@ type ComprehensiveAPIWithoutRaw =
   Vault :> GET :<|>
   Verb 'POST 204 '[JSON] NoContent :<|>
   Verb 'POST 204 '[JSON] Int :<|>
+  Stream 'GET 200 NetstringFraming JSON [Int] :<|>
   WithNamedContext "foo" '[] GET :<|>
   CaptureAll "foo" Int :> GET :<|>
   Summary "foo" :> GET :<|>