+++ servant/test/Servant/LinksSpec.hs
@@ -1,22 +1,25 @@
-{-# LANGUAGE CPP #-}
-{-# LANGUAGE ConstraintKinds #-}
-{-# LANGUAGE DataKinds       #-}
-{-# LANGUAGE PolyKinds       #-}
-{-# LANGUAGE TypeOperators   #-}
+{-# LANGUAGE CPP                 #-}
+{-# LANGUAGE ConstraintKinds     #-}
+{-# LANGUAGE DataKinds           #-}
+{-# LANGUAGE PolyKinds           #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeOperators       #-}
 #if __GLASGOW_HASKELL__ < 709
 {-# OPTIONS_GHC -fcontext-stack=41 #-}
 #endif
-module Servant.Utils.LinksSpec where
+module Servant.LinksSpec where
 
-import           Data.Proxy              (Proxy (..))
-import           Test.Hspec              (Expectation, Spec, describe, it,
-                                          shouldBe)
-import           Data.String             (fromString)
+import           Data.Proxy
+                 (Proxy (..))
+import           Data.String
+                 (fromString)
+import           Test.Hspec
+                 (Expectation, Spec, describe, it, shouldBe)
 
 import           Servant.API
-import           Servant.Utils.Links
-import           Servant.API.Internal.Test.ComprehensiveAPI (comprehensiveAPIWithoutRaw)
+import           Servant.API.Internal.Test.ComprehensiveAPI
+                 (comprehensiveAPIWithoutRaw)
+import           Servant.Links
 
 type TestApi =
   -- Capture and query params
@@ -51,7 +54,7 @@ shouldBeLink link expected =
     toUrlPiece link `shouldBe` fromString expected
 
 spec :: Spec
-spec = describe "Servant.Utils.Links" $ do
+spec = describe "Servant.Links" $ do
     it "generates correct links for capture query params" $ do
         let l1 = Proxy :: Proxy ("hello" :> Capture "name" String :> Delete '[JSON] NoContent)
         apiLink l1 "hi" `shouldBeLink` "hello/hi"